creating dpcpp version of gpairs

+++ native/gpairs/GPU/gpairs.cpp
@@ -7,7 +7,7 @@
 #include <omp.h>
 #include "euro_opt.h"
 
-void call_gpairs( int npoints, tfloat* x1, tfloat* y1, tfloat* z1, tfloat* w1, tfloat* x2,tfloat* y2,tfloat* z2, tfloat* w2,tfloat* rbins,tfloat* results_test) {
+void call_gpairs( size_t npoints, tfloat* x1, tfloat* y1, tfloat* z1, tfloat* w1, tfloat* x2,tfloat* y2,tfloat* z2, tfloat* w2,tfloat* rbins,tfloat* results_test) {
 
   int nbins = DEFAULT_NBINS;
